Add files via upload
With the intended code implemented, I had to learn how a string array works, how to showcase the list of the games index I entered, how to show the list, and how to edit the names to replace and remove them. This code placed a good challenge when adding in the empty slots to the array, adding a game count for no overflow, inputting getline after cin due to character buffer, and code structure with having many brackets for breaking out of the loop. This code would provide valuable information in case I need to look back at something I forgot how to do.