Add files via upload
This had given me trouble when implementing the getters and setters of how I want to showcase the levels or numbers of how I want the human class to extract energy while also regaining energy back from rest. Of course, this was purely trial and error, mainly errors hassling me, but at the same time gaining more better understanding of how this would work with the private and public display. The most satisfying part is when I was able to display the course of the swings and rest to finally let me finish this code.