Summary to Saving Ships
This was something I have to look back at many times as I keep on making the same mistake when inputting the Constructor and Getters and when inputting the object into the file. I revisit this code to what it needed until finally I put the ship object as parameters that will help create and open the Ship to write the name, fuel, and type, even have the error handling in place. While for the others I inputted private members to help with the writeShip To File function.